Android Portion
It contains an android section of our project. However, it could not be uploaded due to large file error (135mb) therefore 'build' folder has been deleted. It will be re-created when the program will run. 
It contains two interfaces, one for Members and another for Employees. Almost all of them are completed, just a little work is left along with integration with the web portion.